[moe refactor] change test model from fake moe model to mixtral moe layer and remove useless test
